% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmTestDS.R
\name{lmTestDS}
\alias{lmTestDS}
\title{Computes a lm}
\usage{
lmTestDS(formula = NULL, data = NULL)
}
\arguments{
\item{formula}{for the lm}

\item{data}{used for lm}
}
\value{
parts required for lm
}
\description{
This function calculates a lm
}
\details{
Test lm
}
\author{
Stefan B., Daniel S.
}
